requirements.txt
numpy==1.22.3
Pillow
opencv-python
matplotlib